feat: Add constructTokenURI function for generating token URIs
This commit adds the constructTokenURI function to the utils.js file. The function takes in parameters for name, image, description, and animation_url and constructs a metadata object. It then sends a POST request to the PUBLIC_IPFS_UPLOADS endpoint with the metadata as the body. The response contains the cid of the uploaded metadata, which is used to generate the token URI. This function enhances the functionality of the application by providing a convenient way to generate token URIs for NFTs.

Note: This commit message follows the conventional commit format and provides a clear summary of the changes made.